Move test support to its own package
