+++ pointfreeco.playground/contents.xcplayground
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='5.0' target-platform='macos'>
+<playground version='5.0' target-platform='macos' executeOnSourceChanges='false'>
     <timeline fileName='timeline.xctimeline'/>
 </playground>
